# docs/openapi.yaml
openapi: 3.1.0
info:
  title: voiceobs
  description: Voice AI observability server
  version: 0.0.2
paths:
...
  /ingest:
    post:
      tags:
      - Spans
      summary: Ingest spans
      description: Ingest one or more spans for analysis.
      operationId: ingest_ingest_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/SpanInput'
              - $ref: '#/components/schemas/SpanBatchInput'
              title: Payload
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid span data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /spans/{span_id}:
    get:
      tags:
      - Spans
      summary: Get span by ID
      description: Get a specific span by its UUID with full details.
      operationId: get_span_spans__span_id__get
      parameters:
      - name: span_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Span Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpanDetailResponse'
        '404':
          description: Span not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error type
        message:
          type: string
          title: Message
          description: Error message
        detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Detail
          description: Additional error details
      type: object
      required:
      - error
      - message
      title: ErrorResponse
      description: 'Response model for errors.


        Returned when an API request fails.'
      example:
        detail: Span with ID 550e8400-e29b-41d4-a716-446655440000 not found
        error: not_found
        message: Resource not found
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IngestResponse:
      properties:
        accepted:
          type: integer
          title: Accepted
          description: Number of spans accepted
        span_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Span Ids
          description: IDs of ingested spans
      type: object
      required:
      - accepted
      - span_ids
      title: IngestResponse
      description: 'Response model for span ingestion.


        Returned after successfully ingesting one or more spans.'
      example:
        accepted: 2
        span_ids:
        - 550e8400-e29b-41d4-a716-446655440000
        - 550e8400-e29b-41d4-a716-446655440001
    SpanBatchInput:
      properties:
        spans:
          items:
            $ref: '#/components/schemas/SpanInput'
          type: array
          title: Spans
          description: List of spans to ingest
      type: object
      required:
      - spans
      title: SpanBatchInput
      description: 'Input model for a batch of spans.


        Use this model to ingest multiple spans in a single request

        for better efficiency.'
      example:
        spans:
        - attributes:
            voice.actor: user
            voice.conversation.id: conv-123
            voice.turn.id: turn-001
          duration_ms: 2500.0
          name: voice.turn
        - attributes:
            voice.conversation.id: conv-123
            voice.stage.type: asr
          duration_ms: 150.5
          name: voice.asr
    SpanDetailResponse:
      properties:
        id:
          type: string
          title: Id
          description: Span UUID
        name:
          type: string
          title: Name
          description: Span name
        start_time:
          anyOf:
          - type: string
          - type: 'null'
          title: Start Time
          description: Start time (ISO 8601)
        end_time:
          anyOf:
          - type: string
          - type: 'null'
          title: End Time
          description: End time (ISO 8601)
        duration_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration Ms
          description: Duration in milliseconds
        attributes:
          additionalProperties: true
          type: object
          title: Attributes
          description: Span attributes
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
          description: OpenTelemetry trace ID
        span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Span Id
          description: OpenTelemetry span ID
        parent_span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Span Id
          description: Parent span ID
      type: object
      required:
      - id
      - name
      title: SpanDetailResponse
      description: Response model for a single span with full details.
      example:
        attributes:
          voice.actor: user
          voice.conversation.id: conv-123
          voice.transcript: Hello, I need help
          voice.turn.id: turn-001
        duration_ms: 2500.0
        end_time: '2024-01-15T10:30:02.500Z'
        id: 550e8400-e29b-41d4-a716-446655440000
        name: voice.turn
        span_id: span-001
        start_time: '2024-01-15T10:30:00Z'
        trace_id: abc123def456
    SpanInput:
      properties:
        name:
          type: string
          title: Name
          description: Span name (e.g., 'voice.turn', 'voice.asr')
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
          description: Span start time (ISO 8601)
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
          description: Span end time (ISO 8601)
        duration_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration Ms
          description: Span duration in milliseconds
        attributes:
          additionalProperties: true
          type: object
          title: Attributes
          description: Span attributes
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
          description: OpenTelemetry trace ID
        span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Span Id
          description: OpenTelemetry span ID
        parent_span_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent Span Id
          description: Parent span ID
      type: object
      required:
      - name
      title: SpanInput
      description: 'Input model for a single span.


        Represents a single observability span from a voice AI pipeline.

        Spans track timing and metadata for individual operations like

        speech recognition (ASR), language model inference (LLM), or

        text-to-speech (TTS).'
      examples:
      - attributes:
          voice.actor: user
          voice.conversation.id: conv-123
          voice.transcript: Hello, how can I help you?
          voice.turn.id: turn-001
          voice.turn.index: 0
        duration_ms: 2500.0
        end_time: '2024-01-15T10:30:02.500Z'
        name: voice.turn
        span_id: span-001
        start_time: '2024-01-15T10:30:00Z'
        trace_id: abc123def456
      - attributes:
          voice.conversation.id: conv-123
          voice.stage.type: asr
        duration_ms: 150.5
        name: voice.asr
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError